+++ Model/Action/Data/OrderRefundActionData.php
@@ -59,8 +59,9 @@ class OrderRefundActionData extends \OxidEsales\PayPalModule\Model\Action\Data\O
     public function getAmount()
     {
         $amount = $this->getRequest()->getRequestParameter('refund_amount');
+        $validDecimalAmount = preg_replace('/,(?=\d{0,2}$)/', '.', $amount) ?? $amount;
 
-        return $amount ? $amount : $this->getPaymentBeingRefunded()->getRemainingRefundAmount();
+        return $validDecimalAmount ? $validDecimalAmount : $this->getPaymentBeingRefunded()->getRemainingRefundAmount();
     }
 
     /**